arm64: dts: sdm845: restore old IPA interconnect providers
This commit:
  392a585583c0 arm64: dts: sdm845: add IPA information
combined changes that were posted as two separate patches.  The two
patches were meant to support both the "old" interconnect provider
definitions and the "new" ones.

The "new" ones were put in place by this commit:
  b303f9f0050b arm64: dts: sdm845: Redefine interconnect provider DT nodes

In order to work in an older environment, we need to change the
interconnect providers back so it's done the old way.

Signed-off-by: Alex Elder <elder@linaro.org>