remoteproc: kill IPA notify code
The IPA code now uses the generic remoteproc SSR notification
mechanism.  This makes the original IPA notification code unused
and unnecessary, so get rid of it.

This is effectively a revert of:
  d7f5f3c89c1a ("remoteproc: add IPA notification to q6v5 driver")

However this was not cherry-picked from this upstream commit:
  2f3ee5e481ce8 ("remoteproc: kill IPA notify code")
because doing that cleanly would also require back-porting some
additional remoteproc commits upstream, which we don't need.

Signed-off-by: Alex Elder <elder@linaro.org>
Change-Id: I2c89cf474e9055935e654a59557b843a4650e1e0